+++ js/src/components/google-combo-account-card/connect-ads/connect-ads-footer.js
@@ -14,17 +14,21 @@ import DisconnectAccount from '.~/components/google-ads-account-card/disconnect-
  *
  * @param {Object} props Props.
  * @param {boolean} props.isConnected Whether the account is connected.
- * @param {Function} props.onCreateNew Callback to create a new account.
+ * @param {Function} props.onCreateNewClick Callback to create a new account.
  * @param {Object} props.restProps Rest props. Passed to AppButton.
  * @return {JSX.Element} Footer component.
  */
-const ConnectAdsFooter = ( { isConnected, onCreateNew, ...restProps } ) => {
+const ConnectAdsFooter = ( {
+	isConnected,
+	onCreateNewClick,
+	...restProps
+} ) => {
 	if ( isConnected ) {
 		return <DisconnectAccount />;
 	}
 
 	return (
-		<AppButton isTertiary onClick={ onCreateNew } { ...restProps }>
+		<AppButton isTertiary onClick={ onCreateNewClick } { ...restProps }>
 			{ __(
 				'Or, create a new Google Ads account',
 				'google-listings-and-ads'
