+++ tests/e2e/utils/pages/setup-mc/step-1-set-up-accounts.js
@@ -380,4 +380,17 @@ export default class SetUpAccountsPage extends MockRequests {
 	getTermsCheckbox() {
 		return this.page.getByLabel( /I accept the terms and conditions/ );
 	}
+
+	/**
+	 * Register the response when connecting an Ads account
+	 *
+	 * @return {Promise<import('@playwright/test').Response>} The response.
+	 */
+	registerAdsAccountsResponse() {
+		return this.page.waitForResponse(
+			( response ) =>
+				response.url().includes( '/gla/ads/accounts' ) &&
+				response.status() === 200
+		);
+	}
 }
